requirements.txt
tensorflow
streamlit
numpy
pillow  # If you're working with images
